wfj-back-develop/src/main/java/com/wangfj/wms/domain/entity/SysParameterType.java
package com.wangfj.wms.domain.entity;

public class SysParameterType {
    private Integer sid;

    private String code;

    private String type;

    public Integer getSid() {
        return sid;
    }

    public void setSid(Integer sid) {
        this.sid = sid;
    }

    public String getCode() {
        return code;
    }

    public void setCode(String code) {
        this.code = code == null ? null : code.trim();
    }

    public String getType() {
        return type;
    }

    public void setType(String type) {
        this.type = type == null ? null : type.trim();
    }
}